Memoize current_frame
Storing a reference to the current frame is more efficient than
computing it each time. This could probably be made even more efficient
by actively setting it in `push_frame` and `pop_frame` instead of
forcing it to be computed for every frame, but I couldn't get that to
work. It's probably something silly I was doing wrong.